Moved scriptDataPropType from the sectionProgressRedux file to a constants file.
