+++ apps/src/templates/sectionProgress/detail/DetailView.jsx
@@ -2,10 +2,8 @@ import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import ReactTooltip from 'react-tooltip';
 import _ from 'lodash';
-import {
-  getCurrentScriptData,
-  scriptDataPropType
-} from '../sectionProgressRedux';
+import {getCurrentScriptData} from '../sectionProgressRedux';
+import {scriptDataPropType} from '../sectionProgressConstants';
 import {sectionDataPropType} from '@cdo/apps/redux/sectionDataRedux';
 import ProgressLegend from '@cdo/apps/templates/progress/ProgressLegend';
 import VirtualizedDetailView from './VirtualizedDetailView';
