+++ apps/src/templates/sectionProgress/standards/StandardsReportCurrentCourseInfo.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import i18n from '@cdo/locale';
 import SafeMarkdown from '@cdo/apps/templates/SafeMarkdown';
-import {scriptDataPropType} from '@cdo/apps/templates/sectionProgress/sectionProgressRedux';
+import {scriptDataPropType} from '@cdo/apps/templates/sectionProgress/sectionProgressConstants';
 import {sectionDataPropType} from '@cdo/apps/redux/sectionDataRedux';
 import color from '@cdo/apps/util/color';
 import {cstaStandardsURL} from './standardsConstants';
