+++ apps/src/templates/sectionProgress/summary/SummaryView.jsx
@@ -2,10 +2,8 @@ import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import ReactTooltip from 'react-tooltip';
 import _ from 'lodash';
-import {
-  getCurrentScriptData,
-  scriptDataPropType
-} from '../sectionProgressRedux';
+import {getCurrentScriptData} from '../sectionProgressRedux';
+import {scriptDataPropType} from '../sectionProgressConstants';
 import SummaryViewLegend from './SummaryViewLegend';
 import VirtualizedSummaryView from './VirtualizedSummaryView';
 import {sectionDataPropType} from '@cdo/apps/redux/sectionDataRedux';
